+++ django_material_demo/polls/apps.py
@@ -2,6 +2,6 @@ from django.apps import AppConfig
 from material.frontend.apps import ModuleMixin
 
 
-class PollsConfig(ModuleMixin, AppConfig):
+class PollsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'polls'
