+++ django_material_demo/polls/urls.py
@@ -1,7 +1,9 @@
 from django.urls import include, path
 from django.views.generic.base import RedirectView
 
-from . import admin_views, views
+from cms.polls import views as cms_views
+
+from . import views
 
 app_name = 'polls'
 urlpatterns = [
@@ -11,16 +13,16 @@ urlpatterns = [
     path('<int:question_id>/vote/', views.vote, name='vote'),
 
     path('admin/', RedirectView.as_view(url='user/', permanent=True), name='index'),
-    path('admin/user/', include(admin_views.UserViewSet().urls)),
+    path('admin/user/', include(cms_views.UserViewSet().urls)),
     path("admin/user/<int:pk>/password/",
-         admin_views.PasswordChangeView.as_view(),
+         cms_views.PasswordChangeView.as_view(),
          name="password_change"),
     path("admin/user/<int:pk>/password/done/",
-         admin_views.PasswordChangeDoneView.as_view(),
+         cms_views.PasswordChangeDoneView.as_view(),
          name="password_change_done",),
 
-    path('admin/file/', include(admin_views.FileViewSet().urls)),
-    path('admin/question/', include(admin_views.QuestionViewSet().urls)),
-    path('admin/vote/', include(admin_views.VoteViewSet().urls)),
-    path('admin/settings', admin_views.SettingsView.as_view(), name='settings'),
+    path('admin/file/', include(cms_views.FileViewSet().urls)),
+    path('admin/question/', include(cms_views.QuestionViewSet().urls)),
+    path('admin/vote/', include(cms_views.VoteViewSet().urls)),
+    path('admin/settings', cms_views.SettingsView.as_view(), name='settings'),
 ]
