+++ django_material_demo/templates/cms/settings.html
