+++ django_material_demo/templates/material/frontend/base_site.html
@@ -28,15 +28,9 @@
 {% block sidenav_items %}
     {{ block.super }}
     <li>
-        <a href="{% url 'polls:settings' %}" class="collapsible-header" data-turbolink="false">
+        <a href="{% url 'settings' %}" class="collapsible-header" data-turbolink="false">
             Settings
             <i class="material-icons">settings</i>
-        </a>
-    </li>
-    <li>
-        <a target="_blank" href="{% url 'polls:home' %}" class="collapsible-header">
-            Open Homepage
-            <i class="material-icons">open_in_new</i>
         </a>
     </li>
 {% endblock %}
